requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-tinymce==4.1.0
pillow==11.1.0
psycopg==3.2.5
psycopg-binary==3.2.5
sqlparse==0.5.3
typing_extensions==4.12.2
uWSGI==2.0.28
